Refactor UCL_GUILayout

+++ UCL_Core_Scripts/UICore/UCL_GUIStyle.cs
@@ -210,7 +210,7 @@ namespace UCL.Core.UI {
         static StyleData s_EditorWindowData = null;
         static StyleData Data => s_Data == null? s_Data = new StyleData() : s_Data;
         static StyleData EditorWindowData => s_EditorWindowData == null ? s_EditorWindowData = new StyleData() : s_EditorWindowData;
-        public static int GetScaledSize(int iSize) => Mathf.RoundToInt(iSize * CurStyleData.Scale);
+        public static int GetScaledSize(float iSize) => Mathf.RoundToInt(iSize * CurStyleData.Scale);
         public static StyleData CurStyleData => IsInEditorWindow ? EditorWindowData : Data;
         /// <summary>
         /// GUIStyle for GUILayout.Box
